move generated 'metrics_schema' pkg to top level
