+++ collector/topicSorter.go
@@ -22,9 +22,9 @@ import (
 	"log"
 	"time"
 
-	"github.com/dcos/dcos-metrics/collector/metrics_schema"
 	"github.com/dcos/dcos-metrics/producers/kafka"
 	"github.com/dcos/dcos-metrics/producers/statsd"
+	"github.com/dcos/dcos-metrics/schema/metrics_schema"
 	"github.com/linkedin/goavro"
 )
